requirements.txt
opencv-python
numpy
PyMuPDF
Pillow
pytesseract
aiofiles
tqdm
psutil
tesserocr